+++ backend/index.js
@@ -1,7 +1,7 @@
 import express from "express";
 import { PORT, mongoDBURL } from "./config.js";
 import mongoose from "mongoose";
-import { Book } from "./models/bookModel.js";
+import booksRoute from './routes/booksRoute.js';
 
 const app = express();
 
@@ -13,102 +13,7 @@ app.get('/', (request, response) => {
     return response.status(234).send('Welcome')
 });
 
-//New book route
-app.post('/books', async (request, response) => {
-    try {
-        if (
-            !request.body.title ||
-            !request.body.author ||
-            !request.body.publishYear
-        ) {
-            return response.status(400).send({
-                message: 'Send all required fields: title, author, publishYear',
-            });
-        }
-        const newBook = {
-            title: request.body.title,
-            author: request.body.author,
-            publishYear: request.body.publishYear,
-        };
-        
-        const book = await Book.create(newBook);
-        
-        return response.status(201).send(book);
-    } catch (error) {
-        console.log(error.message);
-        response.status(500).send({ message: error.message });
-    }
-})
-
-//Get Books router
-app.get('/books', async (request, response) => {
-    try {
-        const books = await Book.find({});
-
-        return response.status(200).json({
-            count: books.length,
-            data: books
-        });
-    } catch (error) {
-        console.log(error.message);
-        response.status(500).send( {message: error.message});
-    }
-});
-
-//Get Book by ID router
-app.get('/books/:id', async (request, response) => {
-    try {
-        const { id } = request.params;
-        const book = await Book.findById(id);
-
-        return response.status(200).json(book);
-    } catch (error) {
-        console.log(error.message);
-        response.status(500).send( {message: error.message});
-    }
-});
-
-//Update book router
-app.put('/books/:id', async (request, response) => {
-    try {
-        if (
-            !request.body.title ||
-            !request.body.author ||
-            !request.body.publishYear
-        ) {
-            return response.status(400).send({
-                message: 'Send all required fields: title, author, publishYear',
-            });
-        }
-
-        const { id } = request.params;
-        const result = await Book.findByIdAndUpdate(id, request.body);
-
-        if (!result) {
-            return response.status(404).json ({ message: 'Book not found'});
-        }
-        return response.status(200).send({ message: 'Book updated succesfully'});
-    } catch (error) {
-        console.log(error.message);
-        response.status(500).send( {message: error.message} );
-    }
-})
-
-//Delete Book router
-app.delete('/books/:id', async (request, response) => {
-    try {
-        const { id } = request.params;
-        const result = await Book.findByIdAndDelete(id);
-
-        if (!result) {
-            return response.status(404).json( {message: 'Book not found'} );
-        }
-        return response.status(200).send( {message:'Book deleted succesfully'});
-    } catch (error) {
-        console.log(error.message);
-        response.status(500).send( {message: error.message} );
-    }
-})
+app.use('/books', booksRoute);
 
 mongoose
     .connect(mongoDBURL)
